Pull validateForm into domFoods.js and make SRP
