+++ lib/domFoods.js
@@ -1,0 +1,20 @@
+export const validateForm = () => {
+  removeValidations()
+  let name = $('#name').val()
+  let calories = $('#calories').val()
+  if (name === "" && calories === "") {
+    $('.form-container').append('<p class="validation">Please enter a food name</p>')
+    $('.form-container').append('<p class="validation">Please enter a calorie amount</p>')
+    return false
+  } else if (name === "") {
+    $('.form-container').append('<p class="validation">Please enter a food name</p>')
+    return false
+  } else if (calories === "") {
+    $('.form-container').append('<p class="validation">Please enter a calorie amount</p>')
+    return false
+  }
+}
+
+const removeValidations = () => {
+  $('.validation').remove()
+}
