+++ lib/index.js
@@ -1,21 +1,5 @@
 import { appendMeals, getEachFoodDiary, filterFoods, removeMealFood } from './domIndex'
-
-const validateForm = () => {
-  removeValidations()
-  let name = $('#name').val()
-  let calories = $('#calories').val()
-  if (name === "" && calories === "") {
-    $('.form-container').append('<p class="validation">Please enter a food name</p>')
-    $('.form-container').append('<p class="validation">Please enter a calorie amount</p>')
-    return false
-  } else if (name === "") {
-    $('.form-container').append('<p class="validation">Please enter a food name</p>')
-    return false
-  } else if (calories === "") {
-    $('.form-container').append('<p class="validation">Please enter a calorie amount</p>')
-    return false
-  }
-}
+import { validateForm } from './domFoods'
 
 const handleResponse = (response) => {
   return response.json()
@@ -31,10 +15,6 @@ const handleResponse = (response) => {
     }
     return json
   })
-}
-
-const removeValidations = () => {
-  $('.validation').remove()
 }
 
 const getAllFoods = (meals) => {
@@ -207,7 +187,6 @@ const countIncrement = () => {
   }
   getAllFoodsDiary()
 }
-
 
 getAllFoodsValidation()
 getAllFoodsDiary()
